+++ pikoci/mysql/migrate/migrations/migrations.go
@@ -16,7 +16,7 @@ type Migration struct {
 // Migrations is the ordered list of all schema migrations. It is defined as
 // a fixed-size array so that the compiler catches ordering conflicts when
 // multiple developers add migrations concurrently.
-var Migrations = [34]Migration{
+var Migrations = [35]Migration{
 	V0Initial,
 	V1ResourceCheckInterval,
 	V2JobsAndBuilds,
@@ -51,4 +51,5 @@ var Migrations = [34]Migration{
 	V31ForEachJobs,
 	V32Workers,
 	V33ClearQueuesColumn,
+	V34WorkerTags,
 }